mgmt-server/requirements.txt
flask[async]
flask-socketio
flask-cors
psycopg2-binary
gevent-websocket
eventlet